Move chessboard components outside of "global" directory
