+++ frontend/src/components/chessboard/Square.tsx
