+++ frontend/src/components/global/chessboards/GameReplayChessboard.tsx
@@ -5,7 +5,7 @@ import {
 } from "../../../interfaces/chessboardGrid";
 import { getSquareClass } from "../../../utils/boardUtils";
 import { capitaliseFirstLetter } from "../../../utils/generalUtils";
-import ChessboardGrid from "../chessboard/ChessboardGrid";
+import ChessboardGrid from "../../chessboard/ChessboardGrid";
 
 function GameReplayChessboard({
 	parsed_fen_string,
